Attempt to use bindings for progress bars => FAIL (works in X.6 only)
git-svn-id: file://localhost/Users/ssp/Developer/svn/EarthAddresser/trunk@442 030910c9-d52c-0410-92f0-b9f1c9c7ad2e

+++ Magic.h
@@ -41,6 +41,8 @@
 
 @interface Magic : NSObject {
 	IBOutlet NSProgressIndicator * progressBar;
+//	double KMLProgress;
+//	double KMLMaximum;
 	BOOL running;
 	
 	NSThread * geocodingThread;
@@ -52,8 +54,8 @@
 	NSString * lookupInfo;
 	NSString * doneMessage;
 	
-	int addressesAreAvailable;
-	int notSearchedCount;
+	NSInteger addressesAreAvailable;
+	NSInteger notSearchedCount;
 	BOOL nonLocatableAddressesExist;
 	BOOL nonLocatableAddressesButtonHidden;
 	IBOutlet NSButton * runGeolocationButton;
